requirements.txt
# pip install -r requirements.txt
openai
azure-identity
azure-keyvault-secrets